+++ bin/tab_completion.sh
@@ -13,7 +13,7 @@ __filedirs_all()
 }
 
 _mde_echo_version() {
-  echo "1.8.4"
+  echo "1.8.5"
 }
 
 _mde() {
@@ -36,7 +36,11 @@ _mde() {
             
               -d) COMPREPLY="0"; return 0 ;;
             
+              --dump-dump-delegate-object) COMPREPLY="0"; return 0 ;;
+            
               --dump-blocks-in-file) COMPREPLY="0"; return 0 ;;
+            
+              --dump-dump-inherited-lines) COMPREPLY="0"; return 0 ;;
             
               --dump-menu-blocks) COMPREPLY="0"; return 0 ;;
             
@@ -88,7 +92,7 @@ _mde() {
   # present matching option names
   #
   if [[ ${cur} == -* ]] ; then
-    opts=("--block-name" "--config" "--debug" "--dump-blocks-in-file" "--dump-menu-blocks" "--dump-selected-block" "--exit" "--filename" "--find" "--help" "--how" "--list-blocks" "--list-count" "--list-default-env" "--list-default-yaml" "--list-docs" "--list-recent-output" "--list-recent-scripts" "--output-execution-summary" "--output-script" "--output-stdout" "--path" "--pwd" "--run-last-script" "--save-executed-script" "--save-execution-output" "--saved-script-folder" "--saved-stdout-folder" "--select-recent-output" "--select-recent-script" "--tab-completions" "--user-must-approve" "--version" "--display-level")
+    opts=("--block-name" "--config" "--debug" "--dump-dump-delegate-object" "--dump-blocks-in-file" "--dump-dump-inherited-lines" "--dump-menu-blocks" "--dump-selected-block" "--exit" "--filename" "--find" "--help" "--how" "--list-blocks" "--list-count" "--list-default-env" "--list-default-yaml" "--list-docs" "--list-recent-output" "--list-recent-scripts" "--output-execution-summary" "--output-script" "--output-stdout" "--path" "--pwd" "--run-last-script" "--save-executed-script" "--save-execution-output" "--saved-script-folder" "--saved-stdout-folder" "--select-recent-output" "--select-recent-script" "--tab-completions" "--user-must-approve" "--version" "--display-level")
     COMPREPLY=( $(compgen -W "$(printf "'%s' " "${opts[@]}")" -- "${cur}") )
 
     return 0
@@ -111,7 +115,11 @@ _mde() {
         
           -d) COMPREPLY=".BOOL."; return 0 ;;
         
+          --dump-dump-delegate-object) COMPREPLY=".BOOL."; return 0 ;;
+        
           --dump-blocks-in-file) COMPREPLY=".BOOL."; return 0 ;;
+        
+          --dump-dump-inherited-lines) COMPREPLY=".BOOL."; return 0 ;;
         
           --dump-menu-blocks) COMPREPLY=".BOOL."; return 0 ;;
         
@@ -166,4 +174,4 @@ _mde() {
 
 complete -o filenames -o nospace -F _mde mde
 # _mde_echo_version
-# echo "Updated: 2023-12-15 23:59:17 UTC"
+# echo "Updated: 2023-12-22 20:00:40 UTC"